+++ views/admin/index/site_access.php
@@ -80,7 +80,7 @@ if (User::getCurrentUser()->getAuthType() == SiteSettings::LOGIN_STD) {
 echo ' ' . \Yii::t('admin', 'siteacc_useraccounts') . '</label>
 </div>';
 
-if ($controller->getParams()->hasWurzelwerk || $controller->getParams()->isSamlActive()) {
+if ($controller->getParams()->isSamlActive()) {
     $method = SiteSettings::LOGIN_WURZELWERK;
     echo '<div class="checkbox wurzelwerk">
   <label>' . Html::checkbox('login[]', in_array($method, $settings->loginMethods), ['value' => $method]) .